LayoutTests/platform/wpe/css3/flexbox/overflow-auto-resizes-correctly-expected.txt

 

FAIL hbox dimensions assert_equals: expected 103 but got 88
FAIL intrinsicHeightBox dimensions assert_equals: expected 103 but got 88
PASS scrollbarSize should not be zero 
FAIL inner dimensions assert_equals: expected 100 but got 85
FAIL vbox dimensions assert_equals: expected 100 but got 85
